Refactor: 날짜 변환 함수 분리 및 적용
