+++ src/components/Calendar/index.tsx
@@ -4,6 +4,7 @@ import 'react-datepicker/dist/react-datepicker.css';
 import { PostDataType } from '@_types/createPost';
 
 import * as S from './styles';
+import formatDate from '@_utils/formatDate';
 
 type CalendarProps = {
   postState: PostDataType;
@@ -24,18 +25,7 @@ const Calendar = ({ postState, setPostState }: CalendarProps) => {
     const formattedDate = format(date || new Date(), "yyyy-MM-dd'T'HH:mm:ss.SSS");
     setPostState((prevData) => ({ ...prevData, eatTime: formattedDate }));
   };
-  function formatDate(inputDate: string) {
-    const date = new Date(inputDate);
-    const year = date.getFullYear();
-    const month = date.getMonth() + 1;
-    const day = date.getDate();
-    const hours = date.getHours();
-    const minutes = date.getMinutes();
 
-    const formattedDate = `${year}년 ${month}월 ${day}일 ${hours}시 ${minutes < 10 ? '0' : ''}${minutes}분`;
-
-    return formattedDate;
-  }
   return (
     <>
       <S.StyledDatePicker
