+++ src/utils/formatDate.ts
@@ -1,0 +1,12 @@
+function formatDate(inputDate: string) {
+  const date = new Date(inputDate);
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+  const hours = date.getHours();
+  const minutes = date.getMinutes();
+
+  return `${year}년 ${month}월 ${day}일 ${hours}시 ${minutes.toString().padStart(2, '0')}분`;
+}
+
+export default formatDate;
